optimized imports, moved FlashMessage to web package
